Day 6

+++ common_lib.h
@@ -3,5 +3,5 @@
 #include "stdbool.h"
 
 void GetLine(char* buffer, char* dest, size_t size);
-int GetNumbersInLine(char* line, int* array);
+int GetNumbersInLine(char* line, long long int* array);
 int IsNumber(char c);
